Split tests into individual classes
