+++ hedera-nft-auction-demo-java-node/src/testIntegration/java/com/hedera/demo/auction/node/test/integration/restapi/GetAuctionsTest.java
@@ -1,0 +1,85 @@
+package com.hedera.demo.auction.node.test.integration.restapi;
+
+import com.google.errorprone.annotations.Var;
+import com.hedera.demo.auction.node.app.SqlConnectionManager;
+import com.hedera.demo.auction.node.app.api.ApiVerticle;
+import com.hedera.demo.auction.node.app.domain.Auction;
+import com.hedera.demo.auction.node.app.repository.AuctionsRepository;
+import com.hedera.demo.auction.node.test.integration.AbstractIntegrationTest;
+import io.vertx.core.DeploymentOptions;
+import io.vertx.core.Vertx;
+import io.vertx.core.json.JsonArray;
+import io.vertx.ext.web.client.WebClient;
+import io.vertx.ext.web.codec.BodyCodec;
+import io.vertx.junit5.VertxExtension;
+import io.vertx.junit5.VertxTestContext;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+import java.sql.SQLException;
+import java.util.concurrent.TimeUnit;
+
+@ExtendWith(VertxExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+public class GetAuctionsTest extends AbstractIntegrationTest {
+
+    private PostgreSQLContainer postgres;
+    private AuctionsRepository auctionsRepository;
+    Vertx vertx;
+
+    @BeforeAll
+    public void beforeAll(VertxTestContext testContext) throws Throwable {
+        this.postgres = new PostgreSQLContainer("postgres:12.6");
+        this.postgres.start();
+        migrate(this.postgres);
+        SqlConnectionManager connectionManager = new SqlConnectionManager(this.postgres.getJdbcUrl(), this.postgres.getUsername(), this.postgres.getPassword());
+
+        this.auctionsRepository = new AuctionsRepository(connectionManager);
+
+        this.vertx = Vertx.vertx();
+
+        DeploymentOptions options = getVerticleDeploymentOptions(this.postgres.getJdbcUrl(), this.postgres.getUsername(), this.postgres.getPassword());
+        this.vertx.deployVerticle(new ApiVerticle(), options, testContext.completing());
+
+        this.webClient = WebClient.create(this.vertx);
+
+        Assertions.assertTrue(testContext.awaitCompletion(5, TimeUnit.SECONDS));
+        if (testContext.failed()) {
+            throw testContext.causeOfFailure();
+        }
+
+    }
+    @AfterAll
+    public void afterAll(VertxTestContext testContext) {
+        this.vertx.close(testContext.completing());
+        this.postgres.close();
+    }
+
+    @Test
+    public void getAuctionsTest(VertxTestContext testContext) throws SQLException {
+        @Var Auction auction = testAuctionObject(1);
+        Auction newAuction1 = auctionsRepository.createComplete(auction);
+        auction = testAuctionObject(2);
+        Auction newAuction2 = auctionsRepository.createComplete(auction);
+
+        webClient.get(9005, "localhost", "/v1/auctions/")
+                .as(BodyCodec.buffer())
+                .send(testContext.succeeding(response -> testContext.verify(() -> {
+                    Assertions.assertNotNull(response);
+                    JsonArray body = new JsonArray(response.body());
+                    Assertions.assertNotNull(body);
+                    Assertions.assertEquals(2, body.size());
+                    verifyAuction(newAuction1, body.getJsonObject(0));
+                    verifyAuction(newAuction2, body.getJsonObject(1));
+
+                    auctionsRepository.deleteAllAuctions();
+                    testContext.completeNow();
+                })));
+    }
+
+}
